agregar logo y pasar el llamado a la api en el home
