+++ components/Navbar/Navbar.tsx
@@ -1,12 +1,16 @@
 import React from "react";
 import Link from "next/link";
+import logo from "../../public/images/BookScapeLogo.png"
 
 const Navbar = () => {
   return (
     <nav>
-      <menu>
-        <Link href="/">Home</Link>
-      </menu>
+       <Link href="/">
+          <img
+            src={logo.src}
+            alt="Logo" 
+          />
+      </Link>
     </nav>
   );
 };
